Feat/alarm : 알람 CRUD 기능 개발

+++ src/main/java/com/dev/classmoa/service/OpinionService.java
@@ -1,8 +1,10 @@
 package com.dev.classmoa.service;
 
+import com.dev.classmoa.domain.entity.Comment;
 import com.dev.classmoa.domain.entity.Lecture;
 import com.dev.classmoa.domain.entity.Member;
 import com.dev.classmoa.domain.entity.Opinion;
+import com.dev.classmoa.domain.repository.CommentRepository;
 import com.dev.classmoa.domain.repository.OpinionRepository;
 
 import lombok.RequiredArgsConstructor;
@@ -13,8 +15,10 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor
 public class OpinionService {
+
     private final OpinionRepository opinionRepository;
     private final LectureService lectureService;
+    private final CommentRepository commentRepository;
 
     public List<Opinion> getOpinions(String lectureId) {
         Lecture lecture = lectureService.getLectureDetail(lectureId);
@@ -34,20 +38,53 @@ public class OpinionService {
     }
 
     // Member 객체가 굳이 필요하지 않을거 같음.
-    public Boolean edit(Opinion opinion){
+    public Boolean edit(Opinion opinion, Member member){
         Opinion savedOpinion = opinionRepository.findById(opinion.getId())
             .orElseThrow(() -> new IllegalArgumentException("not found"));
 
-        return opinionRepository.save(
-            savedOpinion.builder()
-                    .isModified(true)
-                    .content(opinion.getContent())
-                    .build()
-        ).getIsModified();
+        if(savedOpinion.getMember().equals(member)){
+            return opinionRepository.save(
+                savedOpinion.builder()
+                        .isModified(true)
+                        .content(opinion.getContent())
+                        .build()
+            ).getIsModified();
+        }
+        return false;
     }
 
     // Member 객체가 굳이 필요하지 않을거 같음.
-    public void delete(Opinion opinion){
-        opinionRepository.deleteById(opinion.getId());
+    public void delete(Opinion opinion, Member member) {
+        Opinion savedOpinion = opinionRepository.findById(opinion.getId())
+            .orElseThrow(() -> new IllegalArgumentException("not found"));
+        if(savedOpinion.getMember().equals(member)) {
+            opinionRepository.deleteById(opinion.getId());
+        }
+    }
+
+    // 댓글 생성
+    public Long commentCreate(Comment newComment, Member member){
+        return commentRepository.save(
+            newComment.builder()
+                .opinion(newComment.getOpinion())
+                .member(member)
+                .build()
+        ).getId();
+    }
+    public Boolean commentEdit(Comment newComment, Member member){
+        Comment comment = commentRepository.findById(newComment.getId())
+            .orElseThrow(() -> new IllegalArgumentException("not found"));
+        if(comment.getMember().equals(member)){
+            comment.setContent(newComment.getContent());
+        }
+        return false;
+    }
+
+    public void commentDelete(Comment comment, Member member) {
+        Comment savedComment = commentRepository.findById(comment.getId())
+            .orElseThrow(() -> new IllegalArgumentException("not found"));
+        if(savedComment.getMember().equals(member)){
+            commentRepository.deleteById(comment.getId());
+        }
     }
 }
